Moved the unit test to QueryTranslatorTests class, as a more appropriate place for it.
